refactor: inline version package (#129)
On macOS, this project runs into a lot of trouble with both `VERSION` and
`version` existing in the same directory. Basically, only one or the
other is able to exist at any point in time, so it's virtually
impossible to get a clean git setup.

The good news is the version package only really existed to house a few
constants, so moving those to the main package was relatively
straightforward. It did require changing the build flags in the two
places where they are defined, but that didn't cause trouble.

There are a lot more "modernization" opportunities here, but my goal for
the scope of this PR is just to make it possible to work on the project
on macOS without git issues.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>