build: bump dependency versions (#173)
This bumps Go and Vault to their latest versions. For the Vault SDK, the
latest tagged version does not compile against the latest version of the
root module, so the untagged SDK commit corresponding to the tagged root
module release (it's the same commit) is used to guarantee compatibility.

Also fix up a couple syntax issues in the Dockerfile, which I don't
believe will make a difference, but it will make some warnings go away.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>