build: bump vault version (#178)
The bump vault script didn't work, but oddly just grabbing the latest
version of each of the 3 vault modules did. I'm going to keep the script
in for now in case that ends up being helpful, but if we go one or two
more releases like this, maybe we get to manage updates normally again.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>